refactor(web): relocate Main component to shell directory
Move the Main component to the shell directory for better
organization. This change helps in maintaining a clearer structure by
 grouping related components together. Adjusted import paths
 accordingly to reflect the updated location.
